Refactoring of JS and CSS
