+++ js/api/posts/index.mjs
@@ -1,5 +1,6 @@
 export * from "./create.mjs";
+export * from "./delete.mjs";
 export * from "./getPost.mjs";
 export * from "./getPosts.mjs";
 export * from "./update.mjs";
-export * from "./delete.mjs";
+export * from "./userPosts.mjs";
