+++ js/storage/index.mjs
@@ -1,17 +1,5 @@
-export function save(key, value) {
-  localStorage.setItem(key, JSON.stringify(value));
-}
-
-export function load(key) {
-  try {
-    const value = localStorage.getItem(key);
-    return JSON.parse(value);
-  } catch {
-    return null
-  }
-
-}
-
-export function remove(key) {
-  localStorage.removeItem(key);
-}
+export * from "./isAuthenticated.mjs";
+export * from "./load.mjs";
+export * from "./remove.mjs";
+export * from "./requireAuth.mjs";
+export * from "./save.mjs";
