+++ js/storage/isAuthenticated.mjs
@@ -1,0 +1,3 @@
+export function isAuthenticated() {
+  return !!localStorage.getItem("token");
+}
