+++ js/storage/load.mjs
@@ -1,0 +1,8 @@
+export function load(key) {
+  try {
+    const value = localStorage.getItem(key);
+    return JSON.parse(value);
+  } catch {
+    return null;
+  }
+}
