+++ js/storage/remove.mjs
@@ -1,0 +1,3 @@
+export function remove(key) {
+  localStorage.removeItem(key);
+}
