+++ js/storage/requireAuth.mjs
@@ -1,0 +1,7 @@
+import { isAuthenticated } from "./isAuthenticated.mjs";
+
+export function requireAuth() {
+  if (!isAuthenticated()) {
+    window.location.href = "/profile/login/";
+  }
+}
