+++ js/storage/save.mjs
@@ -1,0 +1,3 @@
+export function save(key, value) {
+  localStorage.setItem(key, JSON.stringify(value));
+}
