+++ profile/index.html
@@ -75,7 +75,7 @@
                         src="../images/icon_home.png"
                         class="icon-transform"
                         alt="home icon"
-                        title="Home"
+                        title="Feed"
                     /></a>
                     <a href="#"
                       ><img
@@ -96,7 +96,7 @@
                         src="../images/icon_settings.png"
                         class="icon-transform"
                         alt="settings icon"
-                        title="Settings"
+                        title="Profile Settings"
                     /></a>
                   </div>
                   <div
