+++ src/scss/components.scss
@@ -1,0 +1,46 @@
+.modal-content {
+  width: 90%;
+  max-width: 500px;
+}
+
+.modal-dialog {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  min-height: calc(100% - 3.5rem);
+}
+
+.edit-post-button {
+  width: 105px;
+}
+
+/*---------- Loader/Spinner ----------*/
+
+.loading-indicator {
+  display: none;
+  text-align: center;
+  margin-top: 40px;
+}
+
+.loading-indicator p{
+  color: var(--color-white);
+}
+
+.post-loader p{
+  color: var(--color-black);
+}
+
+.spinner {
+  border: 4px solid rgba(0, 0, 0, 0.1);
+  border-top: 4px solid #761313;
+  border-radius: 50%;
+  width: 80px;
+  height: 80px;
+  animation: spin 1s linear infinite;
+  margin: 0 auto;
+}
+
+@keyframes spin {
+  0% { transform: rotate(0deg); }
+  100% { transform: rotate(360deg); }
+}
