+++ src/scss/index.scss
@@ -53,203 +53,8 @@ $container-max-widths: (
 
 /*---------- My custom CSS ----------*/
 
-.dropdown-toggle::after {
-  content: none;
-}
-
-li {
-  list-style-type: none;
-}
-
-.h1-logo {
-  font-family: Bangers, Arial, Helvetica, sans-serif;
-  font-size: clamp(1.6rem, 5vw, 3rem); 
-}
-
-.welcome-text{
-  font-family: Lora, Arial, Helvetica, sans-serif;
-}
-
-
-.userImage {
-  width: 60px;
-  height: 60px;
-  object-fit: cover;
-}
-
-
-.profile-picture, .profile-avatar {
-  width: 100%;
-  max-width: 120px;
-  height: auto;
-  object-fit: cover;
-  border-radius: 50%;
-  aspect-ratio: 1 / 1;
-}
-
-.profile-avatar{
-  max-width: 100px;
-}
-
-.profile-pic{
-  width: 100%;
-  max-width: 60px;
-  height: auto;
-  object-fit: cover;
-  border-radius: 50%;
-  aspect-ratio: 1 / 1;
-}
-
-.banner-image{
-  width: 100%;
-  max-height: 240px;
-  object-fit: cover;
-}
-
-.post-image-container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin: 0px 20px 20px 20px;
-}
-
-.post-image {
-  max-width: 100%;
-  height: auto;
-}
-
-.icon{
-  max-width: 30px;
-  max-height: 30px;
-}
-
-.icon-transform{
-  max-width: 30px;
-  max-height: 30px;
-/*   transition: transform 0.3s, filter 0.3s; */
-  cursor: pointer;
-}
-
-.icon-xs{
-  max-width: 20px;
-  max-height: 20px;
-}
-
-.icon-transform:hover {
-  transform: scale(1.1);
-}
-
-.profile-name {
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.sr-only {
-  position: absolute;
-  width: 1px;
-  height: 1px;
-  padding: 0;
-  margin: -1px;
-  overflow: hidden;
-  clip: rect(0, 0, 0, 0);
-  border: 0;
-  color: #e6e5e5;
-}
-
-.sticky-nav {
-  position: fixed;
-  top: 0;
-  z-index: 1000;
-  width: 100%;
-}
-
-html,
-body {
-  height: 100%;
-}
-
-body {
-  display: flex;
-  flex-direction: column;
-}
-
-.forFixedNav{
-  margin-top: 108px
-}
-
-main {
-  flex: 1;
-}
-
-.footer {
-  width: 100%;
-  padding: 50px;
-}
-
-.form-control:disabled{
-  opacity: 0.5;
-  cursor: not-allowed;
-}
-
-.modal-dialog {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  min-height: calc(100% - 3.5rem);
-}
-
-.modal-content {
-  width: 90%;
-  max-width: 500px;
-}
-
-.edit-post-button {
-  width: 105px;
-}
-
-.tag {
-  font-weight: bold;
-  color: #761313;
-  margin-right: 8px;
-}
-
-.author-name{
-  font-weight: bold;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  max-width: 100%;
-  display: block;
-}
-
-/*---------- Loader/Spinner ----------*/
-
-.loading-indicator {
-  display: none;
-  text-align: center;
-  margin-top: 40px;
-}
-
-.loading-indicator p{
-  color: var(--color-white);
-}
-
-.post-loader p{
-  color: var(--color-black);
-}
-
-.spinner {
-  border: 4px solid rgba(0, 0, 0, 0.1);
-  border-top: 4px solid #761313;
-  border-radius: 50%;
-  width: 80px;
-  height: 80px;
-  animation: spin 1s linear infinite;
-  margin: 0 auto;
-}
-
-@keyframes spin {
-  0% { transform: rotate(0deg); }
-  100% { transform: rotate(360deg); }
-}
+@import "global";
+@import "ui";
+@import "components";
+@import "media";
+@import "layout";
