+++ src/scss/media.scss
@@ -1,0 +1,66 @@
+.userImage {
+  width: 60px;
+  height: 60px;
+  object-fit: cover;
+}
+
+
+.profile-picture, .profile-avatar {
+  width: 100%;
+  max-width: 120px;
+  height: auto;
+  object-fit: cover;
+  border-radius: 50%;
+  aspect-ratio: 1 / 1;
+}
+
+.profile-avatar{
+  max-width: 100px;
+}
+
+.profile-pic{
+  width: 100%;
+  max-width: 60px;
+  height: auto;
+  object-fit: cover;
+  border-radius: 50%;
+  aspect-ratio: 1 / 1;
+}
+
+.banner-image{
+  width: 100%;
+  max-height: 240px;
+  object-fit: cover;
+}
+
+.post-image-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin: 0px 20px 20px 20px;
+}
+
+.post-image {
+  max-width: 100%;
+  height: auto;
+}
+
+.icon{
+  max-width: 30px;
+  max-height: 30px;
+}
+
+.icon-transform{
+  max-width: 30px;
+  max-height: 30px;
+  cursor: pointer;
+}
+
+.icon-xs{
+  max-width: 20px;
+  max-height: 20px;
+}
+
+.icon-transform:hover {
+  transform: scale(1.1);
+}
