+++ src/scss/ui.scss
@@ -1,0 +1,26 @@
+.dropdown-toggle::after {
+  content: none;
+}
+
+li {
+  list-style-type: none;
+}
+
+.sticky-nav {
+  position: fixed;
+  top: 0;
+  z-index: 1000;
+  width: 100%;
+}
+
+.sr-only {
+  position: absolute;
+  width: 1px;
+  height: 1px;
+  padding: 0;
+  margin: -1px;
+  overflow: hidden;
+  clip: rect(0, 0, 0, 0);
+  border: 0;
+  color: #e6e5e5;
+}
